requirements.txt
requests
beautifulsoup4
googlesearch-python
requests
beautifulsoup4
googlesearch-python
pandas
